+++ props/box_sizing.go
@@ -1,0 +1,12 @@
+package props
+
+type BoxSizing string
+
+const (
+	BoxSizingBorderBox  BoxSizing = "border-box"
+	BoxSizingContentBox BoxSizing = "content-box"
+)
+
+func (b BoxSizing) String() string {
+	return string(b)
+}